requirements.txt
Flask==2.3.2
numpy==1.25
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.1+cpu 
sentence-transformers==2.2.1
pinecone-client==2.2.2
pandas==2.0.3
Flask-Cors==4.0.0
weaviate-client==3.23.0